backend/requirements.txt
Django>=4.2
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers
requests
gunicorn
django-environ
cryptography
PyJWT>=2.0.0
